Added some server fixes and structural changes.
