+++ instanyan/src/components/routing/ProtectedRoute.js
@@ -1,0 +1,22 @@
+import { useSelector } from 'react-redux'
+import { useNavigate, Outlet } from 'react-router-dom'
+
+const ProtectedRoute = () => {
+  const { userInfo } = useSelector((state) => state.user)
+
+  // show unauthorized screen if no user is found in redux store
+  if (!userInfo) {
+    return (
+      <div className='unauthorized'>
+        <h1>Unauthorized :(</h1>
+        <span>
+          <NavLink to='/login'>Login</NavLink> to gain access
+        </span>
+      </div>
+    )
+  }
+
+  // returns child route elements
+  return <Outlet />
+}
+export default ProtectedRoute
