+++ server/nyan-server/config/keys.py
@@ -1,0 +1,1 @@
+JWT_SECRET = 'billybob'
